requirements.txt
python-dotenv==1.0.0
langchain==0.0.352
langsmith==0.0.75
openai==1.6.1
chainlit==1.0.0
qdrant-client==1.7.0
sentence-transformers==2.2.2
ragas==0.0.22
llama-index==0.9.21
jupyter==1.0.0
spacy==3.7.2
tiktoken==0.5.2
pypdf==3.17.4